feat: enable exporting snapshot from statistic page
